--- vortex-array/src/array/transform/mod.rs
@@ -1,15 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: Copyright the Vortex contributors
-
-pub mod context;
-pub mod optimizer;
-pub mod rules;
-#[cfg(test)]
-mod tests;
-
-pub use context::ArrayRuleContext;
-pub use optimizer::ArrayOptimizer;
-pub use rules::AnyArrayParent;
-pub use rules::ArrayParentMatcher;
-pub use rules::ArrayParentReduceRule;
-pub use rules::ArrayReduceRule;
