+++ vortex-array/src/arrays/bool/vtable/operator.rs
@@ -5,19 +5,18 @@ use vortex_compute::filter::Filter;
 use vortex_error::VortexResult;
 use vortex_vector::bool::BoolVector;
 
-use crate::ArrayRef;
-use crate::IntoArray;
-use crate::array::transform::ArrayParentReduceRule;
-use crate::array::transform::ArrayRuleContext;
+use crate::array::optimizer::rules::ArrayParentReduceRule;
 use crate::arrays::BoolArray;
 use crate::arrays::BoolVTable;
 use crate::arrays::MaskedArray;
 use crate::arrays::MaskedVTable;
+use crate::execution::kernel;
 use crate::execution::BatchKernelRef;
 use crate::execution::BindCtx;
-use crate::execution::kernel;
 use crate::vtable::OperatorVTable;
 use crate::vtable::ValidityHelper;
+use crate::ArrayRef;
+use crate::IntoArray;
 
 impl OperatorVTable<BoolVTable> for BoolVTable {
     fn bind(
@@ -54,7 +53,6 @@ impl ArrayParentReduceRule<BoolVTable, MaskedVTable> for BoolMaskedValidityRule 
         array: &BoolArray,
         parent: &MaskedArray,
         _child_idx: usize,
-        _ctx: &ArrayRuleContext,
     ) -> VortexResult<Option<ArrayRef>> {
         // Merge the parent's validity mask into the child's validity
         // TODO(joe): make this lazy
