+++ vortex-array/src/arrays/expr/mod.rs
@@ -5,5 +5,4 @@ mod array;
 pub use array::ExprArray;
 
 mod vtable;
-pub use vtable::ExprOptimizationRule;
 pub use vtable::ExprVTable;
