+++ vortex-array/src/arrays/expr/vtable/mod.rs
@@ -4,20 +4,16 @@
 mod array;
 mod canonical;
 mod operations;
-pub mod operator;
 mod visitor;
 
 use std::fmt::Debug;
 
-pub use operator::ExprOptimizationRule;
 use vortex_buffer::BufferHandle;
 use vortex_dtype::DType;
-use vortex_error::VortexResult;
 use vortex_error::vortex_bail;
+use vortex_error::VortexResult;
 use vortex_vector::Vector;
 
-use crate::Array;
-use crate::ArrayOperator;
 use crate::arrays::expr::ExprArray;
 use crate::execution::ExecutionCtx;
 use crate::expr::Expression;
@@ -28,6 +24,8 @@ use crate::vtable::ArrayVTable;
 use crate::vtable::ArrayVTableExt;
 use crate::vtable::NotSupported;
 use crate::vtable::VTable;
+use crate::Array;
+use crate::ArrayOperator;
 
 vtable!(Expr);
 
@@ -46,7 +44,7 @@ impl VTable for ExprVTable {
     type VisitorVTable = Self;
     type ComputeVTable = NotSupported;
     type EncodeVTable = NotSupported;
-    type OperatorVTable = Self;
+    type OperatorVTable = NotSupported;
 
     fn id(&self) -> ArrayId {
         ArrayId::new_ref("vortex.expr")
