+++ vortex-array/src/arrays/masked/vtable/mod.rs
@@ -10,14 +10,10 @@ mod validity;
 use vortex_buffer::BufferHandle;
 use vortex_compute::mask::MaskValidity;
 use vortex_dtype::DType;
-use vortex_error::VortexResult;
 use vortex_error::vortex_bail;
+use vortex_error::VortexResult;
 use vortex_vector::Vector;
 
-use crate::ArrayBufferVisitor;
-use crate::ArrayChildVisitor;
-use crate::ArrayOperator;
-use crate::EmptyMetadata;
 use crate::arrays::masked::MaskedArray;
 use crate::execution::ExecutionCtx;
 use crate::serde::ArrayChildren;
@@ -30,6 +26,10 @@ use crate::vtable::NotSupported;
 use crate::vtable::VTable;
 use crate::vtable::ValidityVTableFromValidityHelper;
 use crate::vtable::VisitorVTable;
+use crate::ArrayBufferVisitor;
+use crate::ArrayChildVisitor;
+use crate::ArrayOperator;
+use crate::EmptyMetadata;
 
 vtable!(Masked);
 
@@ -119,15 +119,15 @@ mod tests {
     use rstest::rstest;
     use vortex_buffer::ByteBufferMut;
 
-    use crate::ArrayContext;
-    use crate::IntoArray;
-    use crate::MaskedVTable;
     use crate::arrays::MaskedArray;
+    use crate::arrays::MaskedVTable;
     use crate::arrays::PrimitiveArray;
     use crate::serde::ArrayParts;
     use crate::serde::SerializeOptions;
     use crate::validity::Validity;
     use crate::vtable::ArrayVTableExt;
+    use crate::ArrayContext;
+    use crate::IntoArray;
 
     #[rstest]
     #[case(
